Match type for scroll_flag member to fmmdisp function parameter.

+++ trunk/src/linad99/fvar.hpp
@@ -3002,7 +3002,7 @@ class fmm_control
    long ialph;
    long ihflag;
    long ihang;
-   long scroll_flag;
+   int scroll_flag;
    int maxfn_flag;
    int quit_flag;
    double min_improve;
